~ BlockChain package refactoring
