+++ server/blockchain/chain.go
@@ -6,7 +6,8 @@ import (
   // "encoding/base64"
   // "rhymald/mag-delta/funcs"
   // "rhymald/mag-delta/player"
-  // "time"
+  "time"
+  "encoding/base64"
 )
 
 type BlockChain struct {
@@ -43,6 +44,26 @@ type bcIterator struct {
 //   })
 //   return err
 // }
+
+func FindByPrefixes(chain *BlockChain, prefix []byte) [][]byte {
+  var playerList [][]byte
+  chain.Database.View( func(txn *badger.Txn) error {
+    iterator := txn.NewIterator(badger.DefaultIteratorOptions)
+    defer iterator.Close()
+    for iterator.Seek(prefix); iterator.ValidForPrefix(prefix); iterator.Next() {
+      item := iterator.Item()
+      key := item.Key()
+      err := item.Value(func (v []byte) error {
+        playerList = append(playerList, []byte(fmt.Sprintf("%s = %x", key, v)))
+        return nil
+      })
+      if err != nil { return err }
+    }
+    return nil
+  })
+  if len(playerList) == 0 { playerList = append(playerList, prefix) }
+  return playerList
+}
 
 func InitBlockChain(dbPath string) *BlockChain {
   // return &BlockChain{[]*Block{Genesis()}}
@@ -122,4 +143,48 @@ func deeper(iter *bcIterator) *block {
   if err != nil { fmt.Println(err) }
   iter.Current = block.Prev // step back
   return block
+}
+
+func ListBlocks(chain *BlockChain, namespace string) []string {
+  var rows []string
+  var playerIDs []string
+  iter := iterator(chain)
+  depth := 0
+  next := &block{Time: time.Now().UnixNano(), Namespace: namespace}
+  fmt.Println(" ─┼───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────")
+  for i:=0; i<10; i++ {
+    each := deeper(iter)
+    if each.Namespace != next.Namespace { fmt.Printf(" ─┼─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────── \n") }
+    fmt.Printf("  │ \u001b[1m%x\u001b[0m\n", string(each.Hash))
+    fmt.Printf(" ─┼─── %d'", -depth)
+    fmt.Printf("\u001b[1m%s\u001b[0m", each.Namespace)
+    fmt.Printf(" ─── \u001b[1mTime\u001b[0m %d", each.Time)
+    if each.Namespace == "Players[]" {
+      rows = append(rows, fmt.Sprintf("Players[%.8X] = %x", each.Hash, each.Hash))
+      playerIDs = append(playerIDs, fmt.Sprintf("Players[%.8X]", each.Hash))
+    }
+    fmt.Printf(" ─── \u001b[1mGape\u001b[0m %0.3fs.", float64(each.Time-next.Time)/1000000000)
+    fmt.Printf(" ─── \u001b[1mNonce\u001b[0m %d", each.Nonce)
+    fmt.Printf(" ─── \u001b[1mValid\u001b[0m %v\n", validate(newProof(each, Diff[each.Namespace])))
+    decoded, _ := base64.StdEncoding.DecodeString(string(each.Data))
+    fmt.Printf("  │ \u001b[1mData\u001b[0m %s\n", decoded)
+    // fmt.Printf("  │ %x\n", each.Prev)
+    // if each.Namespace != namespace { break }
+    depth--
+    if len(each.Prev) == 0 { break } else { fmt.Printf("  │ %x\n", each.Prev) }
+    next = each
+    // pow := NewProof(each)
+  }
+  // fmt.Println("  │ \n")
+  fmt.Println(" ─┼───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────\n")
+  playerList := FindByPrefixes(chain, []byte("Session[00"))
+  if namespace == "Players[]" {
+    fmt.Println("    ─────────────────── Born blocks: ─────────────────────────────────────────────────────────────────────────────────────────────────────────")
+    for _, link := range rows { fmt.Println(link) }
+    for _, each := range playerList { fmt.Println(string(each)) }
+    fmt.Println("    ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────")
+  }
+  fmt.Println()
+  if len(playerIDs) == 0 { playerIDs = append(playerIDs, "Players[]") }
+  return playerIDs
 }
